refactor(app): :art: Video package refactor of test suite to reduce factory pattern code duplication
